Separate task list into new TaskList component
* Made new TaskList component, which takes the list of user tasks as a prop and sorts displays them accordingly.
* App now displays message if user is not logged in
